+++ src/tls.rs
@@ -789,7 +789,7 @@ impl ServerCertVerifier for IgnoreHostname {
 #[derive(Clone)]
 pub struct TlsInfo {
     pub(crate) peer_certificate: Option<Vec<u8>>,
-    pub(crate) tls_exporter: Option<Vec<u8>>,
+    pub(crate) tls_exporter_channel_binding: Option<Vec<u8>>,
 }
 
 impl TlsInfo {
@@ -799,8 +799,8 @@ impl TlsInfo {
     }
 
     /// Get the [RFC 9266] `tls-exporter` channel binding for this connection.
-    pub fn tls_exporter(&self) -> Option<&[u8]> {
-        self.tls_exporter.as_ref().map(|v| &v[..])
+    pub fn tls_exporter_channel_binding(&self) -> Option<&[u8]> {
+        self.tls_exporter_channel_binding.as_ref().map(|v| &v[..])
     }
 }